User request returns 500
- Remove Not null constraint on users elements after migration

+++ shanoir-ng-users/src/main/java/org/shanoir/ng/user/service/UserServiceImpl.java
@@ -80,11 +80,12 @@ public class UserServiceImpl implements UserService {
 			LOG.error("User with id {} not found", user.getId());
 			throw new EntityNotFoundException(User.class, user.getId());
 		}
-		if (!userDb.isAccountRequestDemand() && !userDb.isExtensionRequestDemand()) {
+		if ((userDb.isAccountRequestDemand() == null || !userDb.isAccountRequestDemand())
+				&& (userDb.isExtensionRequestDemand() == null || !userDb.isExtensionRequestDemand())) {
 			throw new AccountNotOnDemandException(user.getId());
 		}
 		// Confirm and update user
-		if (userDb.isExtensionRequestDemand()) {
+		if (userDb.isExtensionRequestDemand() != null && userDb.isExtensionRequestDemand()) {
 			// Date extension
 			userDb.setExtensionRequestInfo(null);
 			userDb.setExtensionRequestDemand(false);
@@ -121,10 +122,11 @@ public class UserServiceImpl implements UserService {
 		if (user == null) {
 			throw new EntityNotFoundException(User.class, userId);
 		}
-		if (!user.isAccountRequestDemand() && !user.isExtensionRequestDemand()) {
+		if ((user.isAccountRequestDemand() == null || !user.isAccountRequestDemand())
+				&& (user.isExtensionRequestDemand() == null || !user.isExtensionRequestDemand())) {
 			throw new AccountNotOnDemandException(userId);
 		}
-		if (user.isAccountRequestDemand()) {
+		if (user.isAccountRequestDemand() != null && user.isAccountRequestDemand()) {
 			// Remove user
 			userRepository.delete(userId);
 			keycloakClient.deleteUser(user.getKeycloakId());
@@ -292,7 +294,7 @@ public class UserServiceImpl implements UserService {
 	 * @return database user with new values.
 	 */
 	private User updateUserValues(final User userDb, final User user) {
-		userDb.setCanAccessToDicomAssociation(user.isCanAccessToDicomAssociation());
+		userDb.setCanAccessToDicomAssociation(user.isCanAccessToDicomAssociation() != null && user.isCanAccessToDicomAssociation());
 		userDb.setEmail(user.getEmail());
 		userDb.setExpirationDate(user.getExpirationDate());
 		userDb.setFirstName(user.getFirstName());
